# perf: improve performance by using slice