+ 原本的 benchmark 在比較時會一直 push 同一個 element，但這個 push 僅時淺複製，很不合理
+ 改善後的 push, pop 時間複雜度皆為 O(1), get 則為 O(log n)

```
goos: linux
goarch: amd64
pkg: github.com/HazelnutParadise/go-chunkpipe
cpu: QEMU Virtual CPU version 2.5+
                   │    old.txt     │               new2.txt               │
                   │     sec/op     │    sec/op     vs base                │
Push10x10000-8          886.8µ ± 3%   603.8µ ± 18%  -31.91% (p=0.000 n=10)
Push100x1000-8          85.85µ ± 3%   22.58µ ±  1%  -73.69% (p=0.000 n=10)
Push1000x100-8         11.280µ ± 1%   3.098µ ±  1%  -72.53% (p=0.000 n=10)
Push10000x10-8         3831.5n ± 2%   585.1n ±  2%  -84.73% (p=0.000 n=10)
PopEnd10x10000-8       4200.3µ ± 2%   460.9µ ±  1%  -89.03% (p=0.000 n=10)
PopEnd100x1000-8       3956.8µ ± 1%   414.6µ ±  2%  -89.52% (p=0.000 n=10)
PopEnd1000x100-8       3945.3µ ± 2%   416.3µ ±  2%  -89.45% (p=0.000 n=10)
PopEnd10000x10-8       3909.2µ ± 2%   424.6µ ±  1%  -89.14% (p=0.000 n=10)
PopFront10x10000-8     2413.7µ ± 1%   904.6µ ±  5%  -62.52% (p=0.000 n=10)
PopFront100x1000-8     2094.4µ ± 3%   368.5µ ±  8%  -82.40% (p=0.000 n=10)
PopFront1000x100-8     2057.7µ ± 2%   319.0µ ±  1%  -84.50% (p=0.000 n=10)
PopFront10000x10-8     2067.4µ ± 1%   318.5µ ±  5%  -84.59% (p=0.000 n=10)
Get10x10000-8        2059.043m ± 4%   2.667m ±  3%  -99.87% (p=0.000 n=10)
Get100x1000-8         170.576m ± 3%   1.637m ±  3%  -99.04% (p=0.000 n=10)
Get1000x100-8          10.912m ± 3%   1.200m ±  2%  -89.01% (p=0.000 n=10)
Get10000x10-8          1701.6µ ± 3%   780.2µ ±  2%  -54.15% (p=0.000 n=10)
geomean                 2.075m        239.3µ        -88.47%

                   │     old.txt     │                 new2.txt                  │
                   │      B/op       │      B/op       vs base                   │
Push10x10000-8        631.8Ki ± 0%     1436.6Ki ±  0%  +127.39% (p=0.000 n=10)
Push100x1000-8        69.27Ki ± 0%      68.59Ki ±  0%    -0.97% (p=0.000 n=10)
Push1000x100-8       13.016Ki ± 0%      9.094Ki ±  0%   -30.13% (p=0.000 n=10)
Push10000x10-8         7568.0 ± 0%        992.0 ±  0%   -86.89% (p=0.000 n=10)
PopEnd10x10000-8        0.000 ± 0%        0.000 ±  0%         ~ (p=1.000 n=10) ¹
PopEnd100x1000-8        0.000 ± 0%        0.000 ±  0%         ~ (p=1.000 n=10) ¹
PopEnd1000x100-8        0.000 ± 0%        0.000 ±  0%         ~ (p=1.000 n=10) ¹
PopEnd10000x10-8        0.000 ± 0%        0.000 ±  0%         ~ (p=1.000 n=10)
PopFront10x10000-8      0.000 ± 0%        8.000 ± 38%         ? (p=0.000 n=10)
PopFront100x1000-8      0.000 ± 0%        0.000 ±  0%         ~ (p=1.000 n=10) ¹
PopFront1000x100-8      0.000 ± 0%        0.000 ±  0%         ~ (p=1.000 n=10) ¹
PopFront10000x10-8      0.000 ± 0%        0.000 ±  0%         ~ (p=1.000 n=10) ¹
Get10x10000-8           0.000 ± 0%        0.000 ±  0%         ~ (p=1.000 n=10) ¹
Get100x1000-8           0.000 ± 0%        0.000 ±  0%         ~ (p=1.000 n=10) ¹
Get1000x100-8           0.000 ± 0%        0.000 ±  0%         ~ (p=1.000 n=10) ¹
Get10000x10-8           0.000 ± 0%        0.000 ±  0%         ~ (p=1.000 n=10) ¹
geomean                            ²                   ?                       ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                   │     old.txt     │               new2.txt               │
                   │    allocs/op    │ allocs/op   vs base                  │
Push10x10000-8       10024.00 ± 0%     19.00 ± 0%  -99.81% (p=0.000 n=10)
Push100x1000-8        1024.00 ± 0%     11.00 ± 0%  -98.93% (p=0.000 n=10)
Push1000x100-8        124.000 ± 0%     8.000 ± 0%  -93.55% (p=0.000 n=10)
Push10000x10-8         34.000 ± 0%     5.000 ± 0%  -85.29% (p=0.000 n=10)
PopEnd10x10000-8        0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
PopEnd100x1000-8        0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
PopEnd1000x100-8        0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
PopEnd10000x10-8        0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
PopFront10x10000-8      0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
PopFront100x1000-8      0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
PopFront1000x100-8      0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
PopFront10000x10-8      0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
Get10x10000-8           0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
Get100x1000-8           0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
Get1000x100-8           0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
Get10000x10-8           0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
geomean                            ²               -61.95%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```